Remove Oro/UIBundle (merge useful pieces in Pim/UIBundle)
