--- src/Oro/Bundle/UIBundle/.gitignore
@@ -1,10 +1,0 @@
-public_html/bundles/
-app/bootstrap.php.cache
-app/cache/*
-app/logs/*
-build/
-vendor
-composer.phar
-app/config/parameters.yml
-.idea
-.idea/*
