--- src/Oro/Bundle/UIBundle/OroUIBundle.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Oro\Bundle\UIBundle;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class OroUIBundle extends Bundle
-{
-}
