--- src/Oro/Bundle/UIBundle/Resources/config/services.yml
@@ -1,31 +1,0 @@
-parameters:
-    oro_ui.twig.ceil.class:              Oro\Bundle\UIBundle\Twig\CeilExtension
-    oro_ui.twig.extension.class:         Oro\Bundle\UIBundle\Twig\UiExtension
-    oro_ui.twig.md5.class:               Oro\Bundle\UIBundle\Twig\Md5Extension
-    oro_ui.twig.date.class:              Oro\Bundle\UIBundle\Twig\DateExtension
-
-services:
-    oro_ui.twig.ceil_extension:
-        class: '%oro_ui.twig.ceil.class%'
-        tags:
-            - { name: twig.extension }
-
-    oro_ui.twig.ui_extension:
-        class: '%oro_ui.twig.extension.class%'
-        arguments: ['%oro_ui.placeholders%', '%oro_ui.wrap_class%']
-        tags:
-            - { name: twig.extension }
-
-    oro_ui.twig.md5_extension:
-        class: '%oro_ui.twig.md5.class%'
-        arguments:
-            - '@serializer'
-        tags:
-            - { name: twig.extension }
-
-    oro_ui.twig.date_extension:
-        class: '%oro_ui.twig.date.class%'
-        arguments:
-            - '@translator'
-        tags:
-            - { name: twig.extension }
