--- src/Oro/Bundle/UIBundle/Resources/translations/messages.en.yml
@@ -1,2 +1,0 @@
-oro.age: "{0} newborn|]0,Inf[age %count%"
-"Get help": "Get help"
