--- src/Oro/Bundle/UIBundle/Tests/Unit/Fixture/UnitTestBundle.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Oro\Bundle\UIBundle\Tests\Unit\Fixture;
-
-class UnitTestBundle
-{
-}
