--- src/Oro/Bundle/UIBundle/Tests/Unit/Twig/Template/TestJS.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace Oro\Bundle\UIBundle\Tests\Unit\Twig\Template;
-
-use Oro\Bundle\UIBundle\Twig\Template;
-
-class TestJS extends Template
-{
-    public function getTemplateName()
-    {
-        return 'app.js.twig';
-    }
-
-    protected function doDisplay(array $context, array $blocks = [])
-    {
-        echo 'console.log("test")';
-    }
-}
