--- src/Oro/Bundle/UIBundle/composer.json
@@ -1,28 +1,0 @@
-{
-    "name": "oro/ui-bundle",
-    "type": "symfony-bundle",
-    "description": "BAP User Interface Bundle",
-    "keywords": ["BAP", "UI", "User Interface"],
-    "homepage": "https://github.com/laboro/UIBundle",
-    "license": "MIT",
-    "require": {
-        "php": ">=5.3.3",
-        "symfony/framework-bundle": ">=2.1,<2.3-dev",
-        "kriswallsmith/assetic": "1.1.*@dev",
-        "symfony/assetic-bundle": "2.1.*",
-        "leafo/lessphp": "dev-master",
-        "willdurand/expose-translation-bundle": "0.2.*@dev",
-        "oro/requirejs-bundle": "dev-master",
-        "oro/locale-bundle": "dev-master"
-    },
-    "autoload": {
-        "psr-0": { "Oro\\Bundle\\UIBundle": "" }
-    },
-    "minimum-stability": "dev",
-    "target-dir": "Oro/Bundle/UIBundle",
-    "extra": {
-        "branch-alias": {
-            "dev-master": "1.0.0-alpha"
-        }
-    }
-}
