--- src/Pim/Bundle/EnrichBundle/Resources/config/bundles/oro_ui.yml
@@ -1,6 +1,0 @@
-oro_ui:
-    settings:
-        application_name:
-            value: Akeneo
-        application_title:
-            value: Akeneo
