+++ src/Pim/Bundle/EnrichBundle/Resources/config/pim.yml
@@ -12,7 +12,6 @@ imports:
     - { resource: '@PimEnrichBundle/Resources/config/bundles/oro_filter.yml' }
     - { resource: '@PimEnrichBundle/Resources/config/bundles/oro_require_js.yml' }
     - { resource: '@PimEnrichBundle/Resources/config/bundles/oro_translation.yml' }
-    - { resource: '@PimEnrichBundle/Resources/config/bundles/oro_ui.yml' }
     - { resource: '@PimEnrichBundle/Resources/config/bundles/oro_user.yml' }
     - { resource: '@PimEnrichBundle/Resources/config/bundles/twig.yml' }
