+++ src/Pim/Bundle/UIBundle/DependencyInjection/Configuration.php
@@ -23,13 +23,44 @@ class Configuration implements ConfigurationInterface
 
         $rootNode = $treeBuilder
             ->root('pim_ui')
-            ->children()
-                ->booleanNode('loading_message_enabled')
-                    ->defaultTrue()
+                ->children()
+                    ->scalarNode('wrap_class')
+                        ->cannotBeEmpty()
+                        ->defaultValue('block-wrap')
+                        ->end()
+                    ->arrayNode('placeholders_items')
+                        ->useAttributeAsKey('name')
+                        ->prototype('array')
+                        ->children()
+                        ->arrayNode('items')
+                        ->prototype('array')
+                            ->children()
+                                ->booleanNode('remove')->defaultValue(false)->end()
+                                ->scalarNode('placeholder')->end()
+                                ->scalarNode('order')->end()
+                            ->end()
+                        ->end()
+                    ->end()
+                    ->booleanNode('loading_message_enabled')
+                        ->defaultTrue()
+                    ->end()
                 ->end()
             ->end();
 
-        SettingsBuilder::append($rootNode, ['loading_message_enabled' => ['value' => true]]);
+        SettingsBuilder::append(
+            $rootNode,
+            [
+                'application_name' => [
+                    'value' => 'Akeneo',
+                    'type'  => 'scalar'
+                ],
+                'application_title' => [
+                    'value' => 'Akeneo',
+                    'type'  => 'scalar'
+                ],
+                'loading_message_enabled' => ['value' => true]
+            ]
+        );
 
         return $treeBuilder;
     }
