+++ src/Pim/Bundle/UIBundle/Resources/config/twig.yml
@@ -1,9 +1,19 @@
 parameters:
+    pim_ui.twig.extension.class: Pim\Bundle\UIBundle\Twig\UiExtension
     pim_ui.twig.style_extension.class: Pim\Bundle\UIBundle\Twig\StyleExtension
     pim_ui.twig.loading_message_extension.class: Pim\Bundle\UIBundle\Twig\LoadingMessageExtension
     pim_ui.loading_messages_file: '@@PimUIBundle/Resources/config/loading_messages.txt'
+    pim_ui.twig.ceil.class: Pim\Bundle\UIBundle\Twig\CeilExtension
+    pim_ui.twig.md5.class:  Pim\Bundle\UIBundle\Twig\Md5Extension
+    pim_ui.twig.date.class: Pim\Bundle\UIBundle\Twig\DateExtension
 
 services:
+    pim_ui.twig.ui_extension:
+        class: '%pim_ui.twig.extension.class%'
+        arguments: ['%pim_ui.placeholders%', '%pim_ui.wrap_class%']
+        tags:
+            - { name: twig.extension }
+
     pim_ui.twig.style_extension:
         class: '%pim_ui.twig.style_extension.class%'
         public: false
@@ -16,5 +26,24 @@ services:
             - '@file_locator'
             - '%pim_ui.loading_messages_file%'
         public: false
+        tags:
+            - { name: twig.extension }
+
+    pim_ui.twig.ceil_extension:
+        class: '%pim_ui.twig.ceil.class%'
+        tags:
+            - { name: twig.extension }
+
+    pim_ui.twig.md5_extension:
+        class: '%pim_ui.twig.md5.class%'
+        arguments:
+            - '@serializer'
+        tags:
+            - { name: twig.extension }
+
+    pim_ui.twig.date_extension:
+        class: '%pim_ui.twig.date.class%'
+        arguments:
+            - '@translator'
         tags:
             - { name: twig.extension }
