+++ src/Pim/Bundle/UIBundle/Resources/views/Default/logo.html.twig
@@ -1,5 +1,5 @@
 <h1 class="AknHeader-logo">
-    <a href="{{ path('oro_default') }}" title="{{ oro_config_value('oro_ui.application_title') }}">
-        <img class="AknHeader-logoImage" src="{{ asset('/bundles/pimui/images/logo.png') }}" alt="{{ oro_config_value('oro_ui.application_name') }}">
+    <a href="{{ path('oro_default') }}" title="{{ oro_config_value('pim_ui.application_title') }}">
+        <img class="AknHeader-logoImage" src="{{ asset('/bundles/pimui/images/logo.png') }}" alt="{{ oro_config_value('pim_ui.application_name') }}">
     </a>
 </h1>
