+++ src/Pim/Bundle/UIBundle/Resources/views/actions/update.html.twig
@@ -1,5 +1,5 @@
 {% extends bap.layout%}
-{% import 'OroUIBundle::macros.html.twig' as UI %}
+{% import 'PimUIBundle::macros.html.twig' as UI %}
 {% import 'PimUIBundle:Default:page_elements.html.twig' as elements %}
 
 {% block page_container %}
