+++ src/Pim/Bundle/UIBundle/Resources/views/requirejs.config.js.twig
