+++ src/Pim/Bundle/UIBundle/Twig/Md5Extension.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Oro\Bundle\UIBundle\Twig;
+namespace Pim\Bundle\UIBundle\Twig;
 
 use JMS\Serializer\Exception\RuntimeException;
 use JMS\Serializer\Serializer;
