+++ src/Pim/Bundle/UserBundle/Resources/views/Group/update.html.twig
@@ -1,4 +1,4 @@
-{% extends 'OroUIBundle:actions:update.html.twig' %}
+{% extends 'PimUIBundle:actions:update.html.twig' %}
 {% import 'PimDataGridBundle::macros.html.twig' as dataGrid %}
 
 {% set entityId = form.vars.value.id %}
