+++ src/Pim/Bundle/UserBundle/Resources/views/Role/update.html.twig
@@ -1,6 +1,6 @@
 {% extends 'PimEnrichBundle::layout.html.twig' %}
 {% import 'PimDataGridBundle::macros.html.twig' as dataGrid %}
-{% import 'OroUIBundle::macros.html.twig' as UI %}
+{% import 'PimUIBundle::macros.html.twig' as UI %}
 
 {% oro_title_set({params : { "%role%": form.vars.value.label } }) %}
